Update references for files that were moved in creation of unit test directory in tests
